move ILM client objects from protocol to client
